Fixing a bug in the Bici integral
This integral is what we used to call dV_0/dPsi/V_0 (I need to come up with a better name for this). The way it was programmed, it only worked for the asymptotic value at x=-\infty, because for some reason I used the asymptotic value of Cd in the denominator. This did not affect the results where only the asymptotic values were plotted (e.g., Fig 3 in the paper), but became necessary in the last figure of the paper. Good that I caught it.